refactor(domain): modularize di container into 8 focused registration modules
Split the monolithic 645-line initializeContainer() function into 8
focused module files under packages/core/src/infrastructure/di/modules/:

- database.module.ts: DB connection, migrations, instance token
- repositories.module.ts: ISettingsRepository, IFeatureRepository, etc.
- services.module.ts: Version, Worktree, Git, IDE, Daemon, Deployment, etc.
- agent-infrastructure.module.ts: executor factory/provider, runner, registry
- notifications.module.ts: notification bus, desktop notifier, service
- use-cases.module.ts: all use case singletons + session repositories
- interactive.module.ts: interactive session repos, service, use cases
- web-tokens.module.ts: string-token aliases for Turbopack web routes

Pure refactor — no behavior changes, all bindings preserved.
